third_party/WebKit/LayoutTests/fast/forms/text/text-dir-auto-with-placeholder.html

<!DOCTYPE html>
<meta charset="utf-8">
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<input dir="auto" placeholder="LTR placeholder string" value="&#x05E2;">
<script>
test(() => {
    assert_equals(getComputedStyle(document.querySelector('input'), '').direction, 'rtl');
}, 'Placeholder\'s direction shouldn\'t affect INPUT direction.');
</script>
</body>
